This is my group's Machine Learning project!
Hospital readmissions are of utmost importance to assess healthcare quality, especially 
regarding chronic illnesses such as diabetes. In light of this, it is logical to investigate the challenge 
presented by readmissions, as it can have a great impact not only on patient care but also on 
healthcare costs.
The objective of this project is to leverage machine learning algorithms for the development 
of predictive models capable of accurately forecasting hospital readmissions for diabetic patients. To 
this end, our work was twofold: firstly, we implemented a binary classification model that predicts 
whether a patient will be readmitted within 30 days of discharge and, secondly, we developed a 
multiclass classification model that categorizes the readmission timeframe as “NO”, “<30days” and 
“>30days”. The binary classifier seeks to aid healthcare providers in implementing specific preventative 
measures, whilst the multiclass classifier provides detailed insights into varying levels of patient risk, 
enabling more customized post-discharge care.